requirements.txt
aniso8601==9.0.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
click==8.0.4
Flask==2.0.1
Flask-Cors==3.0.10
flask-restx==0.5.0
idna==2.10
importlib-resources==5.4.0
itsdangerous==2.1.0
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.1.0
pycparser==2.21
PyJWT==2.3.0
pymssql==2.2.2
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dotenv==0.19.0
pytz==2021.3
requests==2.23.0
six==1.16.0
urllib3==1.25.11
Werkzeug==2.0.1
zipp==3.7.0
